backend/requirements.txt
docker==6.0.0
requests==2.26.0
django==2.2.19
djangorestframework==3.12.4
PyJWT==2.1.0
django-filter==2.4
gunicorn==20.0.4
Pillow==8.3.1
sorl-thumbnail==12.7.0
djoser
drf-extra-fields
pango
weasyprint
django-cors-headers
python-dotenv==0.19.2
django-colorfield==0.7.2